requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.2
anyio==4.8.0
astroid==3.3.11
async-timeout==4.0.3
attrs==25.1.0
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.1
botocore==1.34.162
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
click==8.1.8
coloredlogs==15.0.1
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
django-picklefield==3.2
djangorestframework_simplejwt==5.4.0
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
future==0.18.3
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
groq==0.18.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.0
langchain-text-splitters==0.0.2
langsmith==0.1.147
lxml==5.3.1
macholib 
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
parsel==1.10.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
posthog==3.14.1
propcache==0.2.1
Protego==0.4.0
proto-plus==1.26.0
protobuf==5.29.3
pulsar-client==3.6.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
pylint==3.3.9
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.11.6
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
Scrapy==2.12.0
service-identity==24.2.0
shellingham==1.5.4
six 
tenacity==8.5.0
threadpoolctl==3.5.0
tldextract==5.1.3
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
Twisted==24.11.0
twython==3.9.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
django-cors-headers==4.9.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
django-filter
gunicorn
psycopg2-binary
whitenoise
dj-database-url
